Rename celery file to avoid circular imports
